+++ arc/command/option/custom/member.py
@@ -5,19 +5,20 @@ import typing as t
 import attr
 import hikari
 
-from arc.abc.option import CommandOptionBase, OptionParams, OptionType
+from arc.abc.option import ConverterOption, OptionParams, OptionType
+from arc.errors import OptionConverterFailureError
 from arc.internal.types import ClientT
 
 if t.TYPE_CHECKING:
     import typing_extensions as te
 
 
-__all__ = ("UserOption", "UserParams")
+__all__ = ("MemberOption", "MemberParams")
 
 
 @t.final
-class UserParams(OptionParams[hikari.User]):
-    """The parameters for a user option.
+class MemberParams(OptionParams[hikari.InteractionMember]):
+    """The parameters for a member option.
 
     Parameters
     ----------
@@ -38,23 +39,30 @@ class UserParams(OptionParams[hikari.User]):
 
 
 @attr.define(slots=True, kw_only=True)
-class UserOption(CommandOptionBase[hikari.User, ClientT, UserParams]):
-    """A slash command option that represents a user.
+class MemberOption(ConverterOption[hikari.InteractionMember, ClientT, MemberParams, hikari.User]):
+    """A slash command option that represents a member.
 
     ??? hint
         To add an option of this type to your command, add an argument to your command function with the following type hint:
         ```py
-        opt_name: arc.Option[hikari.User, UserParams(...)]
+        opt_name: arc.Option[hikari.Member, MemberParams(...)]
         ```
     """
 
     @property
     def option_type(self) -> OptionType:
-        return OptionType.USER
+        return OptionType.MEMBER
+
+    def _convert_value(self, value: hikari.User) -> hikari.InteractionMember:
+        if isinstance(value, hikari.InteractionMember):
+            return value
+        raise OptionConverterFailureError(
+            self, value, f"Option '{self.name}' expected an InteractionMember, got {value!r}."
+        )
 
     @classmethod
     def _from_params(
-        cls, *, name: str, arg_name: str, is_required: bool, params: UserParams, **kwargs: t.Any
+        cls, *, name: str, arg_name: str, is_required: bool, params: MemberParams, **kwargs: t.Any
     ) -> te.Self:
         return cls(
             name=name,