+++ tests/test_sigparse.py
@@ -9,22 +9,24 @@ from arc.internal.sigparse import BASE_CHANNEL_TYPE_MAP, CHANNEL_TYPES_MAPPING, 
 
 async def correct_command(
     ctx: arc.GatewayContext,
-    a: arc.Option[int, arc.IntParams(description="foo", min=10)],
-    b: arc.Option[str, arc.StrParams(description="bar", min_length=100)],
-    c: arc.Option[float | None, arc.FloatParams(description="baz", max=50.0)],
-    d: arc.Option[hikari.GuildTextChannel | hikari.GuildNewsChannel | None, arc.ChannelParams(description="qux")],
-    e: arc.Option[hikari.GuildChannel | None, arc.ChannelParams(description="quux")],
-    f: arc.Option[hikari.Role | hikari.User | None, arc.MentionableParams(description="quuz")] = None,
-    g: arc.Option[hikari.Attachment | None, arc.AttachmentParams(description="among us")] = None,
-    h: arc.Option[bool, arc.BoolParams(description="among us")] = False,
-    i: arc.Option[str, arc.StrParams(description="foo")] = "among us",
+    a: arc.Option[int, arc.IntParams("foo", min=10)],
+    b: arc.Option[str, arc.StrParams("bar", min_length=100)],
+    c: arc.Option[float | None, arc.FloatParams("baz", max=50.0)],
+    d: arc.Option[hikari.GuildTextChannel | hikari.GuildNewsChannel | None, arc.ChannelParams("qux")],
+    e: arc.Option[hikari.GuildChannel | None, arc.ChannelParams("quux")],
+    f: arc.Option[hikari.Role | hikari.User | None, arc.MentionableParams("quuz")] = None,
+    g: arc.Option[hikari.Attachment | None, arc.AttachmentParams("among us")] = None,
+    h: arc.Option[bool, arc.BoolParams("among us")] = False,
+    i: arc.Option[str, arc.StrParams("foo")] = "among us",
+    j: arc.Option[hikari.Member | None, arc.MemberParams("among us")] = None,
+    k: arc.Option[hikari.Color, arc.ColorParams("foo", name="color")] = hikari.Color(0x000000),
 ) -> None:
     pass
 
 
 def test_correct_command() -> None:
     options = parse_command_signature(correct_command)
-    assert len(options) == 9
+    assert len(options) == 11
 
     assert isinstance(options["a"], arc.command.IntOption)
     assert options["a"].name == "a"
@@ -90,6 +92,17 @@ def test_correct_command() -> None:
     assert options["i"].name == "i"
     assert options["i"].description == "foo"
     assert not options["i"].is_required
+
+    assert isinstance(options["j"], arc.command.MemberOption)
+    assert options["j"].name == "j"
+    assert options["j"].description == "among us"
+    assert not options["j"].is_required
+
+    assert isinstance(options["color"], arc.command.ColorOption)
+    assert options["color"].name == "color"
+    assert options["color"].arg_name == "k"
+    assert options["color"].description == "foo"
+    assert not options["color"].is_required
 
 
 async def wrong_params_type(